+++ Modules/Vendors/app/Http/Controllers/VendorsController.php
@@ -13,11 +13,15 @@ class VendorsController extends Controller
 {
     public function index(): JsonResponse
     {
+        $this->authorize('viewAny', VendorProfile::class);
+
         return response()->json(VendorProfile::query()->with('user')->latest()->paginate());
     }
 
     public function store(StoreVendorRequest $request): JsonResponse
     {
+        $this->authorize('create', VendorProfile::class);
+
         $vendor = VendorProfile::create($request->validated());
 
         return response()->json($vendor->load('user'), Response::HTTP_CREATED);
@@ -25,11 +29,15 @@ class VendorsController extends Controller
 
     public function show(VendorProfile $vendor): JsonResponse
     {
+        $this->authorize('view', $vendor);
+
         return response()->json($vendor->load('user'));
     }
 
     public function update(UpdateVendorRequest $request, VendorProfile $vendor): JsonResponse
     {
+        $this->authorize('update', $vendor);
+
         $vendor->update($request->validated());
 
         return response()->json($vendor->refresh()->load('user'));
@@ -37,6 +45,8 @@ class VendorsController extends Controller
 
     public function destroy(VendorProfile $vendor): Response
     {
+        $this->authorize('delete', $vendor);
+
         $vendor->delete();
 
         return response()->noContent();
